Add server environment modules
* [ ] needs https://github.com/OCA/server-env/pull/13

It adds the possibility to configure the values either with:
* `server_environment_files` with files per environment
* `SERVER_ENV_CONFIG` multiline environment variable with the config (https://github.com/OCA/server-env/pull/12)

Thanks to the new features of https://github.com/OCA/server-env/pull/13,
it allows as well to have only some fields in configuration files, and
configure the rest on the UI with <field>_env_default fields (they are
not shown, handled by "inverse" methods). This modules also integrates
the password handling with the "keychain" addon, if you don't set any
password in the configuration files/environment variable, then the
password is read/stored from the keychain account.